feat(build): migrated from testacular to karma
This entailed several changes to many files, mostly a grep/sed affair.
For easier management, we also migrated the local build task for
Testacular/Karma to the NPM package `grunt-karma`.

Node was updated to 0.10 and the Travis CI config to run with Karma. Due
to the changes in the Node version, the Grunt tasks were updated to
their latest version.