feat(*): far better Bower integration
Bower is now the official delivery channel for Twitter Bootstrap and UI
Bootstrap and is an integral part of how ngBoilerplate works. Several
files and directories were renamed for this change to make sense.

Notably, src/components is no longer for vendor code, but for
drag-and-drop reusable code created by the user. Vendor code retrieved
through Bower is stored in the vendor directory instead.

AngularJS files are now in a folder called build to accommodate the
above
change, and so the Grunt tasks are now in a folder called grunt-tasks.

The documentation has been changed to reflect these changes.

As this is the major change for v0.2, the package version was updated.

Closes #23. Closes #22. Closes #12.